requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.2
zipp==3.12.0
flask-login
flask-wtf
py-algorand-sdk
python-dotenv
Flask-HTTPAuth
